+++ R/breaks.R
@@ -17,13 +17,22 @@ y_numeric_breaks <- function(y_var_vctr,
     y_var_vctr <- c(-y_var_vctr, y_var_vctr)
   }
   if (y_zero == TRUE) {
-    y_breaks <- pretty(c(0, y_var_vctr), n = y_pretty_n)
-    if(y_trans %in% c("log10", "log")) y_breaks <- c(1, y_breaks[y_breaks > 1])
-    y_limits <- c(min(y_breaks), max(y_breaks))
+      y_breaks <- pretty(c(0, y_var_vctr), n = y_pretty_n)
+      y_limits <- c(min(y_breaks), max(y_breaks))
+      if(y_trans == "log10" | y_trans == "log") {
+        y_breaks[1] <- 0.1 
+        y_limits[1] <- 0.1
+      }
   }
   else if (y_zero == FALSE) {
     y_breaks <- pretty(y_var_vctr, n = y_pretty_n)
+    y_limits <- c(min(y_breaks), max(y_breaks))
+    if(y_trans == "log10" | y_trans == "log") {
+      if(y_breaks[1] == 0) y_breaks[1] <- 0.1 
+      if(y_limits[1] == 0) y_limits[1] <- 0.1 
+    }
   }
+
   return(y_breaks)
 }
 
@@ -49,11 +58,19 @@ x_numeric_breaks <- function(x_var_vctr,
   }
   if (x_zero == TRUE) {
     x_breaks <- pretty(c(0, x_var_vctr), n = x_pretty_n)
-    if(x_trans %in% c("log10", "log")) x_breaks <- c(1, x_breaks[x_breaks > 1])
     x_limits <- c(min(x_breaks), max(x_breaks))
+    if(x_trans == "log10" | x_trans == "log") {
+      x_breaks[1] <- 0.1 
+      x_limits[1] <- 0.1
+    }
   }
   else if (x_zero == FALSE) {
     x_breaks <- pretty(x_var_vctr, n = x_pretty_n)
+    x_limits <- c(min(x_breaks), max(x_breaks))
+    if(x_trans == "log10" | x_trans == "log") {
+      if(x_breaks[1] == 0) x_breaks[1] <- 0.1 
+      if(x_limits[1] == 0) x_limits[1] <- 0.1 
+    }
   }
   if(mobile == TRUE) {
     x_breaks <- c(min(x_breaks), max(x_breaks))