+++ R/ggplot_line.R
@@ -239,7 +239,7 @@ ggplot_line <- function(data,
 #' @param y_zero_line TRUE or FALSE whether to add a zero reference line to the y axis. TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE. 
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. Not applicable where mobile equals TRUE.
 #' @param col_labels Adjust the  colour scale labels through a vector.
-#' @param col_labels_ncol The number of columns in the legend. Defaults to 1.
+#' @param col_labels_ncol The number of columns in the legend. 
 #' @param col_labels_nrow The number of rows in the legend. 
 #' @param col_na TRUE or FALSE of whether to show NA values of the colour variable. Defaults to TRUE.
 #' @param col_title Colour title string for the legend. Defaults to NULL.
@@ -697,7 +697,7 @@ ggplot_line_facet <-
 #' @param y_zero TRUE or FALSE whether the minimum of the y scale is zero. Defaults to FALSE.
 #' @param y_zero_line TRUE or FALSE whether to add a zero reference line to the y axis. TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE. 
 #' @param col_labels Adjust the  colour scale labels through a vector.
-#' @param col_labels_ncol The number of columns in the legend. Defaults to 1.
+#' @param col_labels_ncol The number of columns in the legend. 
 #' @param col_labels_nrow The number of rows in the legend.
 #' @param col_na TRUE or FALSE of whether to show NA values of the colour variable. Defaults to TRUE.
 #' @param col_title Colour title string for the legend. Defaults to NULL.